+++ web/documentserver-example/csharp-mvc/Helpers/Users.cs
@@ -95,7 +95,7 @@ namespace OnlineEditorsExampleMVC.Helpers
                     descr_user_1,
                     true,
                     true,
-                    new Goback()
+                    new Goback(null, false)
                 ),
             new User(
                     "uid-2",
@@ -135,7 +135,7 @@ namespace OnlineEditorsExampleMVC.Helpers
                     descr_user_3,
                     false,
                     false,
-                    new Goback(null,false)
+                    null
                 ),
             new User(
                     "uid-0",