+++ web/documentserver-example/java-spring/src/main/java/com/onlyoffice/integration/ExampleData.java
@@ -102,7 +102,7 @@ public class ExampleData {
         userService.createUser("John Smith", "smith@example.com", descriptionUserFirst,
                 "", List.of(FilterState.NULL.toString()), List.of(FilterState.NULL.toString()),
                 List.of(FilterState.NULL.toString()), List.of(FilterState.NULL.toString()),
-                List.of(FilterState.NULL.toString()), null, true, true, true, new Goback());
+                List.of(FilterState.NULL.toString()), null, true, true, true, new Goback(null, false));
 
         // create user 2 with the specified parameters
         userService.createUser("Mark Pottato", "pottato@example.com", descriptionUserSecond,
@@ -113,7 +113,7 @@ public class ExampleData {
         // create user 3 with the specified parameters
         userService.createUser("Hamish Mitchell", null, descriptionUserThird,
                 "group-3", List.of("group-2"), List.of("group-2", "group-3"), List.of("group-2"),
-                new ArrayList<>(), List.of("group-2"), false, true, true, false, new Goback(null, false));
+                new ArrayList<>(), List.of("group-2"), false, true, true, false, null);
 
         // create user 0 with the specified parameters
         userService.createUser("Anonymous", null, descriptionUserZero, "",