+++ web/documentserver-example/nodejs/helpers/users.js
@@ -31,6 +31,7 @@ class User {
     templates,
     avatar,
     goback,
+    close,
   ) {
     this.id = id;
     this.name = name;
@@ -45,6 +46,7 @@ class User {
     this.templates = templates;
     this.avatar = avatar;
     this.goback = goback;
+    this.close = close;
   }
 }
 
@@ -103,7 +105,22 @@ const descrUser0 = [
 ];
 
 const users = [
-  new User('uid-1', 'John Smith', 'smith@example.com', null, null, {}, null, null, [], descrUser1, true, true, {}),
+  new User(
+    'uid-1',
+    'John Smith',
+    'smith@example.com',
+    null,
+    null,
+    {},
+    null,
+    null,
+    [],
+    descrUser1,
+    true,
+    true,
+    { blank: false },
+    { visible: false },
+  ),
   new User(
     'uid-2',
     'Mark Pottato',
@@ -122,6 +139,7 @@ const users = [
     false,
     true,
     { text: 'Go to Documents' },
+    {},
   ),
   new User(
     'uid-3',
@@ -140,9 +158,10 @@ const users = [
     descrUser3,
     false,
     false,
-    { blank: false },
+    null,
+    {},
   ),
-  new User('uid-0', null, null, null, null, {}, [], null, ['protect'], descrUser0, false, false, null),
+  new User('uid-0', null, null, null, null, {}, [], null, ['protect'], descrUser0, false, false, null, null),
 ];
 
 // get a list of all the users
@@ -180,10 +199,7 @@ users.getUsersForProtect = function getUsersForProtect(id) {
   const result = [];
   this.forEach((user) => {
     if (user.id !== id && user.name != null) {
-      result.push({
-        id: user.id,
-        name: user.name,
-      });
+      result.push(user);
     }
   });
   return result;