+++ web/documentserver-example/python/src/utils/users.py
@@ -95,7 +95,7 @@ descr_user_0 = [
 USERS = [
     User('uid-1', 'John Smith', 'smith@example.com',
          '', None, {}, None,
-         None, [], descr_user_1, True, True, {}),
+         None, [], descr_user_1, True, True, {'blank': False}),
     User('uid-2', 'Mark Pottato', 'pottato@example.com',
          'group-2', ['group-2', ''], {
              'view': "",
@@ -111,7 +111,7 @@ USERS = [
              'remove': []
          }, ['group-2'],
          False, ["copy", "download", "print"], descr_user_3, False, False,
-         {'blank': False}),
+         None),
     User('uid-0', None, None,
          '', None, {}, [],
          None, ["protect"], descr_user_0, False, False, None)